# INSTA-LITE/insta-lite-back/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  comment.Comment:
    properties:
      commentId:
        type: string
      content:
        type: string
      createdAt:
        type: string
      likesCount:
        type: integer
      parentCommentId:
        type: string
      postId:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  comment.CommentCreateRequest:
    properties:
      content:
        maxLength: 2000
        type: string
      parentCommentId:
        type: string
    required:
    - content
    type: object
  comment.CommentUpdateRequest:
    properties:
      content:
        maxLength: 2000
        type: string
    required:
    - content
    type: object
  comment.PaginatedCommentResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/comment.Comment'
        type: array
      hasMore:
        type: boolean
      nextCursor:
        type: string
    type: object
  commentlike.CommentLike:
    properties:
      commentId:
        type: string
      createdAt:
        type: string
      reactionType:
        $ref: '#/definitions/like.ReactionType'
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  commentlike.reactionRequest:
    properties:
      reactionType:
        $ref: '#/definitions/like.ReactionType'
    required:
    - reactionType
    type: object
...
  like.Like:
    properties:
      created_at:
        type: string
      postID:
        type: string
      reactionType:
        $ref: '#/definitions/like.ReactionType'
      updated_at:
        type: string
      userID:
        type: string
    type: object
  like.LikeRequest:
    properties:
      reactionType:
        $ref: '#/definitions/like.ReactionType'
    required:
    - reactionType
    type: object
  like.ReactionType:
    enum:
    - like
    - love
    - laugh
    type: string
    x-enum-varnames:
    - ReactionLike
    - ReactionLove
    - ReactionLaugh
...
host: localhost:8080
info:
  contact:
    email: huguescodeur@gmail.com
    name: Hugues Codeur
  description: API REST pour une application de type Instagram (posts, likes, commentaires,
    follow).
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Insta-Lite API
  version: "1.0"
paths:
...
  /comments/{commentID}/likes:
    delete:
      description: Retire la réaction de l'utilisateur connecté sur un commentaire
      parameters:
      - description: UUID du commentaire
        in: path
        name: commentID
        required: true
        type: string
      responses:
        "204":
          description: Like supprimé
        "400":
          description: ID incorrect
          schema:
            type: string
        "404":
          description: Like non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Supprimer un like sur un commentaire
      tags:
      - Comment Likes
    get:
      description: Retourne tous les likes d'un commentaire
      parameters:
      - description: UUID du commentaire
        in: path
        name: commentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/commentlike.CommentLike'
            type: array
        "400":
          description: ID incorrect
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Likes d'un commentaire
      tags:
      - Comment Likes
    patch:
      consumes:
      - application/json
      description: Change le type de réaction de l'utilisateur connecté sur un commentaire
      parameters:
      - description: UUID du commentaire
        in: path
        name: commentID
        required: true
        type: string
      - description: Nouvelle réaction
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commentlike.reactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commentlike.CommentLike'
        "400":
          description: Type invalide
          schema:
            type: string
        "404":
          description: Like non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Modifier une réaction sur un commentaire
      tags:
      - Comment Likes
    post:
      consumes:
      - application/json
      description: Ajoute une réaction (like/love/laugh) à un commentaire
      parameters:
      - description: UUID du commentaire
        in: path
        name: commentID
        required: true
        type: string
      - description: Type de réaction
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commentlike.reactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/commentlike.CommentLike'
        "400":
          description: Type de réaction invalide
          schema:
            type: string
        "409":
          description: Déjà liké
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Liker un commentaire
      tags:
      - Comment Likes
  /comments/{commentID}/likes/me:
    get:
      description: Retourne le like de l'utilisateur connecté sur un commentaire
      parameters:
      - description: UUID du commentaire
        in: path
        name: commentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commentlike.CommentLike'
        "400":
          description: ID incorrect
          schema:
            type: string
        "404":
          description: Like non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Mon like sur un commentaire
      tags:
      - Comment Likes
  /comments/{id}:
    delete:
      description: Supprime un commentaire appartenant à l'utilisateur connecté
      parameters:
      - description: UUID du commentaire
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Commentaire supprimé
        "400":
          description: ID incorrect
          schema:
            type: string
        "404":
          description: Commentaire non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Supprimer un commentaire
      tags:
      - Comments
    patch:
      consumes:
      - application/json
      description: Met à jour le contenu d'un commentaire appartenant à l'utilisateur
        connecté
      parameters:
      - description: UUID du commentaire
        in: path
        name: id
        required: true
        type: string
      - description: Nouveau contenu
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/comment.CommentUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.Comment'
        "400":
          description: Format invalide
          schema:
            type: string
        "404":
          description: Commentaire non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Modifier un commentaire
      tags:
      - Comments
...
  /posts/{postID}/comments:
    get:
      description: Retourne les commentaires racines d'un post, paginés par curseur
      parameters:
      - description: UUID du post
        in: path
        name: postID
        required: true
        type: string
      - description: Curseur de pagination
        in: query
        name: cursor
        type: string
      - description: 'Nombre de résultats (défaut: 10, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.PaginatedCommentResponse'
        "400":
          description: ID incorrect
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Commentaires d'un post
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Ajoute un commentaire (ou une réponse) sur un post. Passer parentCommentId
        pour une réponse.
      parameters:
      - description: UUID du post
        in: path
        name: postID
        required: true
        type: string
      - description: Contenu du commentaire
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/comment.CommentCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/comment.Comment'
        "400":
          description: Format invalide ou commentaire parent ne correspond pas au
            post
          schema:
            type: string
        "404":
          description: Post ou commentaire parent non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Créer un commentaire
      tags:
      - Comments
  /posts/{postID}/likes:
    delete:
      description: Retire la réaction de l'utilisateur connecté sur un post
      parameters:
      - description: UUID du post
        in: path
        name: postID
        required: true
        type: string
      responses:
        "204":
          description: Like supprimé
        "400":
          description: ID invalide
          schema:
            type: string
        "404":
          description: Like non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Supprimer un like sur un post
      tags:
      - Post Likes
    get:
      description: Retourne tous les likes d'un post
      parameters:
      - description: UUID du post
        in: path
        name: postID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/like.Like'
            type: array
        "400":
          description: ID invalide
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Likes d'un post
      tags:
      - Post Likes
    patch:
      consumes:
      - application/json
      description: Change le type de réaction (like/love/laugh) de l'utilisateur connecté
        sur un post
      parameters:
      - description: UUID du post
        in: path
        name: postID
        required: true
        type: string
      - description: Nouvelle réaction
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/like.LikeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/like.Like'
        "400":
          description: ID ou type de réaction invalide
          schema:
            type: string
        "404":
          description: Like non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Modifier une réaction sur un post
      tags:
      - Post Likes
    post:
      consumes:
      - application/json
      description: Ajoute une réaction (like/love/laugh) à un post
      parameters:
      - description: UUID du post
        in: path
        name: postID
        required: true
        type: string
      - description: Type de réaction
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/like.LikeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/like.Like'
        "400":
          description: Type de réaction invalide
          schema:
            type: string
        "404":
          description: Post non trouvé
          schema:
            type: string
        "409":
          description: Déjà liké
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Liker un post
      tags:
      - Post Likes
  /posts/{postID}/likes/me:
    get:
      description: Retourne le like de l'utilisateur connecté sur un post
      parameters:
      - description: UUID du post
        in: path
        name: postID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/like.Like'
        "400":
          description: ID invalide
          schema:
            type: string
        "404":
          description: Like non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Mon like sur un post
      tags:
      - Post Likes
...
securityDefinitions:
  BearerAuth:
    description: 'Entrez le token JWT avec le préfixe "Bearer ". Exemple: "Bearer
      eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"